Move input_batch and block_table to runner folder
